allow override of 'isModern' (#935)
* allow override of 'isModern'

Add an option to let users decide to not use geoJSON, even if the
underlying service supports it. Occasionally handy when ArcGIS Server
has geoJSON geometry problems.

* Explicitly set isModern properly

Keep the default for “isModern” to be optimistically true, to preserve
back compatibility. Fix the test to allow metadata to disable geoJSON.

* Remove default value for 'isModern'

Fix tests for ‘isModern’ to not require a default value; update
‘isModern’ as part of the tests for geoJSON support, allowing for a
user override.

* Kill useless test for undefined

Explicit test for false is sufficient.